chore: Migrate `fastify` tests to `node:test`
